+++ feature/aiplanner/src/main/kotlin/com/miruni/feature/aiplanner/domain/model/PlanPreview.kt
@@ -1,10 +1,10 @@
 package com.miruni.feature.aiplanner.domain.model
 
 data class PlanPreview(
-    val planId: Long, // 상위 일정 ID
+    val planId: Int, // 상위 일정 ID
     val title: String, // 상위 일정 제목
-    val doneCount: Long, // 세부 일정 중 완료 개수
-    val totalCount: Long, // 세부 일정 전체 개수
-    val progressRate: Long, // 진행률 (%)
+    val doneCount: Int, // 세부 일정 중 완료 개수
+    val totalCount: Int, // 세부 일정 전체 개수
+    val progressRate: Int, // 진행률 (%)
     val isDone: Boolean, // 일정 전체에 대한 완료 여부
 )
